Load svg inline instead of separate file
