+++ packages/example/src/components/Account/Account.css
@@ -60,7 +60,7 @@
   align-items: center;
 }
 
-.Logo-img {
+.Logo-container svg {
   height: 40px;
 }
