+++ packages/example/src/components/Connect/Connect.tsx
@@ -1,6 +1,6 @@
 import React, { useCallback, useState } from 'react';
-import ConnectIcon from "./image/connect.svg"
-import MetaMaskIcon from "./image/MetaMask.svg"
+import { ReactComponent as ConnectIcon } from "./image/connect.svg"
+import { ReactComponent as  MetaMaskIcon } from "./image/MetaMask.svg"
 import ArrowRight from "../Icons/ArrowRight"
 import Modal from "./Modal";
 import { connect } from "../../lib/snap";
@@ -25,7 +25,7 @@ const Connect = ({open, onConnected}: ConnectProps) => {
 
   return (
     <Modal open={open}>
-      <img src={ConnectIcon} alt="MetaMaskFlask" className="Connect-flask-icon"/>
+      <ConnectIcon className="Connect-flask-icon" />
       <h2>Connect to MetaMask Bitcoin Snap</h2>
       <p className="Connect-install">If you do not have Bitcoin Snap installed, you will be prompted to do so.</p>
       <a 
@@ -38,7 +38,7 @@ const Connect = ({open, onConnected}: ConnectProps) => {
         <ArrowRight />
       </a>
       <button className="Connect-button" disabled={isConnecting} onClick={connectMetaMask}>
-        <img src={MetaMaskIcon} alt="Install"/>
+        <MetaMaskIcon />
         <span>Connect MetaMask</span>
       </button>
     </Modal>
