+++ packages/example/src/components/Connect/Install.tsx
@@ -1,12 +1,12 @@
 import React from 'react';
-import FlaskIcon from "./image/MetaMaskFlask.svg"
-import MetaMaskIcon from "./image/MetaMask.svg"
+import { ReactComponent as FlaskIcon } from "./image/MetaMaskFlask.svg"
+import { ReactComponent as MetaMaskIcon } from "./image/MetaMask.svg"
 import Modal from "./Modal";
 
 const Install = ({open}: {open: boolean}) => {
   return (
     <Modal open={open}>
-      <img src={FlaskIcon} alt="MetaMaskFlask" className="Connect-flask-icon"/>
+      <FlaskIcon className="Connect-flask-icon"/>
       <h2>Install MetaMask Flask</h2>
       <p className="Connect-install">You need to install the MetaMask Flask extension in order to use Bitcoin Snap.</p>
       <p className="Connect-flask">Please disabled the regular MetaMask extension when using flask version</p>
@@ -16,7 +16,7 @@ const Install = ({open}: {open: boolean}) => {
         target="_blank"
         rel="noopener noreferrer"
       >
-        <img src={MetaMaskIcon} alt="Install"/>
+        <MetaMaskIcon />
         <span>Install MetaMask</span>
       </a>
     </Modal>
