+++ packages/example/src/components/Menu/index.css
@@ -61,7 +61,7 @@
   transition: 0.25s;
 }
 
-.Menu-item img {
+.Menu-item svg {
   margin: 6px 8px;
 }
